move tests to src folder so I can more easilly rsync this folder into other codebases
